pdf_converter/resources/mineru_source/MinerU/projects/web_api/requirements.txt
uvicorn>=0.25.0
fastapi>=0.110.0
python-multipart>=0.0.7
boto3>=1.28.43
Brotli>=1.1.0
click>=8.1.7
fast-langdetect>=0.2.3,<0.3.0
loguru>=0.6.0
numpy>=1.21.6
pydantic>=2.7.2,<2.11
PyMuPDF>=1.24.9,<1.25.0
scikit-learn>=1.0.2
torch>=2.2.2
torchvision
transformers>=4.49.0,!=4.51.0,<5.0.0
pdfminer.six==20231228
tqdm>=4.67.1 